refactor(structured-logs): extract json and access log parsers
